nrfx:mdk: Move section starts within sections they mark
If a symbols that makrs a secion start is put before the actual
section it may happen that linker aligns the section to different
boundary than the symbol is aligner. That may cause the section
starts later than symbol points.

Example: In source code we have a uint64_t global variable that
is not initalized. The variable will be put into .bss. Linker
can align the section to 8 byte boundary due to the variable.
Then __bss_start will point to memory 4 bytes before actual
.bss start. During initialization startup code will zero memory
before .bss and skip last 4 bytes. Some data could be overwritten
and some data could be non-zero. The issue occurs in runtime,
in bad way...

Signed-off-by: Piotr Pryga <pryga.piotr@outlook.com>